Add CandidateConcreteTypeResolvers for Jackson
