+++ fixture-monkey-jackson/src/main/java/com/navercorp/fixturemonkey/jackson/generator/PropertyJsonSubTypesObjectPropertyGenerator.java
@@ -18,36 +18,31 @@
 
 package com.navercorp.fixturemonkey.jackson.generator;
 
-import static com.navercorp.fixturemonkey.jackson.property.JacksonAnnotations.getJacksonAnnotation;
-import static com.navercorp.fixturemonkey.jackson.property.JacksonAnnotations.getRandomJsonSubType;
-
-import java.lang.annotation.Annotation;
-import java.lang.reflect.AnnotatedType;
-import java.lang.reflect.Type;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import javax.annotation.Nullable;
-
 import org.apiguardian.api.API;
 import org.apiguardian.api.API.Status;
-
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
 
 import com.navercorp.fixturemonkey.api.generator.ObjectProperty;
 import com.navercorp.fixturemonkey.api.generator.ObjectPropertyGenerator;
 import com.navercorp.fixturemonkey.api.generator.ObjectPropertyGeneratorContext;
 import com.navercorp.fixturemonkey.api.property.Property;
 import com.navercorp.fixturemonkey.api.property.PropertyGenerator;
-import com.navercorp.fixturemonkey.api.property.PropertyUtils;
-import com.navercorp.fixturemonkey.api.type.Types;
+import com.navercorp.fixturemonkey.jackson.property.PropertyJsonSubTypesConcreteTypeResolver;
 
 @API(since = "0.4.2", status = Status.MAINTAINED)
+@Deprecated
+/**
+ * It is deprecated.
+ * Use {@link PropertyJsonSubTypesConcreteTypeResolver} instaed.
+ */
 public final class PropertyJsonSubTypesObjectPropertyGenerator implements ObjectPropertyGenerator {
 	public static final PropertyJsonSubTypesObjectPropertyGenerator INSTANCE =
 		new PropertyJsonSubTypesObjectPropertyGenerator();
+
+	private static final PropertyJsonSubTypesConcreteTypeResolver DELEGATE =
+		new PropertyJsonSubTypesConcreteTypeResolver();
 
 	@Override
 	public ObjectProperty generate(ObjectPropertyGeneratorContext context) {
@@ -56,54 +51,17 @@ public final class PropertyJsonSubTypesObjectPropertyGenerator implements Object
 			.generate(context);
 		PropertyGenerator propertyGenerator = context.getPropertyGenerator();
 
-		JsonSubTypes jsonSubTypes = getJacksonAnnotation(property, JsonSubTypes.class);
-		if (jsonSubTypes == null) {
-			throw new IllegalArgumentException("@JsonSubTypes is not found " + property.getType().getTypeName());
-		}
+		Property actualConcreteProperty = DELEGATE.resolve(property).get(0);
 
-		Class<?> type = getRandomJsonSubType(jsonSubTypes);
-		AnnotatedType annotatedType = Types.generateAnnotatedTypeWithoutAnnotation(type);
 		List<Property> childProperties =
-			propertyGenerator.generateChildProperties(PropertyUtils.toProperty(annotatedType));
+			propertyGenerator.generateChildProperties(actualConcreteProperty);
 
-		JsonTypeInfo jsonTypeInfo = getJacksonAnnotation(property, JsonTypeInfo.class);
-		List<Annotation> annotations = new ArrayList<>(property.getAnnotations());
-		annotations.add(jsonTypeInfo);
-
-		Property actualProperty = new Property() {
-			@Override
-			public Type getType() {
-				return type;
-			}
-
-			@Override
-			public AnnotatedType getAnnotatedType() {
-				return annotatedType;
-			}
-
-			@Nullable
-			@Override
-			public String getName() {
-				return property.getName();
-			}
-
-			@Override
-			public List<Annotation> getAnnotations() {
-				return Collections.unmodifiableList(annotations);
-			}
-
-			@Nullable
-			@Override
-			public Object getValue(Object instance) {
-				return property.getValue(instance);
-			}
-		};
 		return new ObjectProperty(
-			actualProperty,
+			actualConcreteProperty,
 			context.getPropertyNameResolver(),
 			nullInject,
 			context.getElementIndex(),
-			Collections.singletonMap(actualProperty, childProperties)
+			Collections.singletonMap(actualConcreteProperty, childProperties)
 		);
 	}
 }
